# Faster video recording