--- a/gym/envs/classic_control/rendering.py
+++ b/gym/envs/classic_control/rendering.py
@@ -60,7 +60,7 @@ class Viewer(object):
     def add_onetime(self, geom):
         self.onetime_geoms.append(geom)
 
-    def render(self, return_rgb_array):
+    def render(self, return_rgb_array=False):
         glClearColor(1,1,1,1)
         self.window.clear()
         self.window.switch_to()