Running environments with video recording turned on is painfully slow. Let's investigate. In `random_agent.py`:

```python
        for j in range(max_steps):
            action = agent.act(ob, reward, done)
            import time
            t1 = time.time()
            ob, reward, done, _ = env.step(action)
            t2 = time.time()
            print "step %0.2fms" % ((t2-t1)*1000)
```

Now let's try `python examples/agents/random_agent.py CartPole-v0`

```
step 101.11ms
step 99.19ms
step 98.89ms
step 100.18ms
step 99.97ms
step 99.45ms
step 100.70ms
step 99.77ms
step 100.09ms
```

Why 100ms? My computer, MacBook Pro 13" is connected to 4K monitor. Refresh rate is 30 Hz. Pyglet window is created with vsync=True. Every `viewer.window.flip()` waits for vsync. Then `viewer.get_array()` contains another two `flip()`. 1000ms/30*3 = 100ms. Not only this waits for another two vsync events, but is also jitters image a frame back and it's visible.

It doesn't affect resulting encoded video, but the process doesn't look pretty. 

This PR returns array without extra `flip()`s.

```
step 26.68ms
step 39.91ms
step 26.64ms
step 38.33ms
```

Effect is identical for `CartPole`, `Acrobot`, `MountainCar`, `Pendulum`, `LunarLander`, `BipedalWalker`.

Atari and MuJoCo needs to be checked.