@tlbtlbtlb 2016-05-23
I prefer this much smaller change to mujoco-py that accomplishes the same thing.
https://github.com/openai/mujoco-py/commit/bda1ac4dc34501eb1cd3518b179e165140a42e0e