@olegklimov 2016-05-23
Well `vsync=False` is even smaller, but it will not do the same thing: jittering picture in the window will remain. Solution (I'm certainly not saying I have the best possible solution) should involve removing unnecessary `flip()`s, one way or another.