@jietang 2016-06-01
@tlbtlbtlb @olegklimov, what's the status here? is there a use case that requires vsync=False?