@tlbtlbtlb 2016-06-01
Oleg's solution seems good to me, let's merge. It makes a huge difference in non-interactive rendering performance.