@olegklimov 2016-06-02
I guess @tlbtlbtlb didn't mean `vsync=False` makes a difference in non-interactive setting. But it also true. To set `vsync=False` in batch mode we need some indicator it's a batch mode, in constructor. Pyglet doesn't let to change vsync on the fly.

Thinking of it, we have abstract fabric for environments, right?

```
env = gym.make('CartPole-v0', batch_mode=True)
```

Then any window opened by that environment should have vsync=False. Shouldn't be hard (I don't have a patch, maybe later).