@jietang 2016-06-06
@olegklimov I'd rather not add arguments to gym.make(). Perhaps we could add a configure() method you could call on the env to turn on batch mode? e.g.

```
env = gym.make('CartPole-v0')
env.configure(batch_mode=True)
```